image preview and light box 2

+++ app/Http/Controllers/PhotoUploadController.php
@@ -14,7 +14,7 @@ class PhotoUploadController extends Controller
         // $manager = new ImageManager(new Driver());
         // $manager->read($file)->resize($width,$height)->toWebp(75)->save(public_path($path).$name);
         $extension = $file->getClientOriginalExtension();
-        $image_name = $name. '.' .$extension;
+        $image_name = $name . '.' . $extension;
         $file->move(public_path($path), $image_name);
 
         return $image_name;
@@ -22,9 +22,11 @@ class PhotoUploadController extends Controller
 
     public static function imageUnlink($path, $name): void
     {
-        $image_path = public_path($path) . $name;
-        if (file_exists($image_path)) {
-            unlink($image_path);
+        if ($name) {
+            $image_path = public_path($path) . $name;
+            if (file_exists($image_path)) {
+                unlink($image_path);
+            }
         }
     }
 }
